Last modifications, 03/10/2017
Fixed error with empty input sets. The problem was in the fill_in_by_out function, i only checked simple structures to see if they were empty (so i could include them in the 'inicio' set of the individual. Maybe some of the mutation sets could cause this error too, but until now they seem to be ok